requirements.txt
-r requirements-dev.txt

django-widget-tweaks
django-allauth
Pillow
gunicorn
pycopg2
